requirements.txt
discord.py
python-dotenv
wavelink
flask
flask-discord